requirements.txt
fastapi
uvicorn
transformers
pdf2image
Pillow
torch
python-multipart
jinja2
